+++ Source/WebKit2/UIProcess/CoordinatedGraphics/LayerTreeCoordinatorProxy.cpp
@@ -66,9 +66,9 @@ void LayerTreeCoordinatorProxy::updateTileForLayer(int layerID, int tileID, cons
 {
     RefPtr<ShareableSurface> surface;
 #if USE(GRAPHICS_SURFACE)
-    int token = updateInfo.surfaceHandle.graphicsSurfaceToken();
+    uint64_t token = updateInfo.surfaceHandle.graphicsSurfaceToken();
     if (token) {
-        HashMap<uint32_t, RefPtr<ShareableSurface> >::iterator it = m_surfaces.find(token);
+        HashMap<uint64_t, RefPtr<ShareableSurface> >::iterator it = m_surfaces.find(token);
         if (it == m_surfaces.end()) {
             surface = ShareableSurface::create(updateInfo.surfaceHandle);
             m_surfaces.add(token, surface);
